gui/lib/lvgl: define default macros with everything disabled
Process file as if #include "file" appeared as the first line of the
primary source file.

If multiple ‘-include’ options are given, the files are included in the
order they appear on the command line.

'-imacros' is exactly like ‘-include’, except that any output produced
by scanning file is thrown away. Macros it defines remain defined. This
allows you to acquire all the macros from a header without also
processing its declarations.