gui/lib/lvgl: split defaults between lvgl and drivers
Also remove -imacros from 'CPPFLAGS' and enforce developer to include the
defaults in the end of its project 'lv_conf.h' and 'lv_drv_conf.h' if
necessary. It allows for more simple configuration. As -imacros was
included first, the only was to override a configuration entry was to
use '-D' flags, which is not super practical.

Defines in -'D' flags should be use to generate a different binary 'on
the fly' via command line variables or depending of the TARGET in
Makefiles.